+++ esp-hal/src/gpio/any_pin.rs
@@ -15,7 +15,10 @@ impl Inverted {
     }
 }
 
-/// Generic pin wrapper for pins which can be Output or Input.
+/// A type-erased GPIO pin, with additional configuration options.
+///
+/// Note that accessing unsupported pin functions (e.g. trying to use an
+/// input-only pin as output) will panic.
 pub struct AnyPin<'d> {
     pin: ErasedPin,
     inverted: Inverted,
@@ -25,9 +28,7 @@ pub struct AnyPin<'d> {
 impl<'d> AnyPin<'d> {
     /// Create wrapper for the given pin.
     #[inline]
-    pub fn new<P: OutputPin + InputPin + CreateErasedPin>(
-        pin: impl crate::peripheral::Peripheral<P = P> + 'd,
-    ) -> Self {
+    pub fn new<P: CreateErasedPin>(pin: impl crate::peripheral::Peripheral<P = P> + 'd) -> Self {
         crate::into_ref!(pin);
         let pin = pin.erased_pin(private::Internal);
 
@@ -195,88 +196,6 @@ impl<'d> InputPin for AnyPin<'d> {
                 force_via_gpio_mux,
                 private::Internal,
             );
-        }
-    }
-}
-
-/// Generic pin wrapper for pins which can only be Input.
-pub struct AnyInputOnlyPin<'d> {
-    pin: ErasedPin,
-    inverted: Inverted,
-    _phantom: PhantomData<&'d ()>,
-}
-
-impl<'d> AnyInputOnlyPin<'d> {
-    /// Create wrapper for the given pin.
-    #[inline]
-    pub fn new<P: InputPin + CreateErasedPin>(
-        pin: impl crate::peripheral::Peripheral<P = P> + 'd,
-    ) -> Self {
-        crate::into_ref!(pin);
-        let pin = pin.erased_pin(private::Internal);
-
-        Self {
-            pin,
-            inverted: Inverted::NonInverted,
-            _phantom: PhantomData,
-        }
-    }
-}
-
-impl<'d> crate::peripheral::Peripheral for AnyInputOnlyPin<'d> {
-    type P = Self;
-
-    unsafe fn clone_unchecked(&mut self) -> Self::P {
-        Self {
-            pin: unsafe { self.pin.clone_unchecked() },
-            inverted: self.inverted,
-            _phantom: PhantomData,
-        }
-    }
-}
-
-impl<'d> private::Sealed for AnyInputOnlyPin<'d> {}
-
-impl<'d> Pin for AnyInputOnlyPin<'d> {
-    delegate::delegate! {
-        to self.pin {
-            fn number(&self, _internal: private::Internal) -> u8;
-            fn sleep_mode(&mut self, on: bool, _internal: private::Internal);
-            fn set_alternate_function(&mut self, alternate: AlternateFunction, _internal: private::Internal);
-            fn is_listening(&self, _internal: private::Internal) -> bool;
-            fn listen_with_options(
-                &mut self,
-                event: Event,
-                int_enable: bool,
-                nmi_enable: bool,
-                wake_up_from_light_sleep: bool,
-                _internal: private::Internal,
-            );
-            fn unlisten(&mut self, _internal: private::Internal);
-            fn is_interrupt_set(&self, _internal: private::Internal) -> bool;
-            fn clear_interrupt(&mut self, _internal: private::Internal);
-            fn wakeup_enable(&mut self, enable: bool, event: WakeEvent, _internal: private::Internal);
-        }
-    }
-}
-
-impl<'d> InputPin for AnyInputOnlyPin<'d> {
-    delegate::delegate! {
-        to self.pin {
-            fn init_input(&self, pull_down: bool, pull_up: bool, _internal: private::Internal);
-            fn set_to_input(&mut self, _internal: private::Internal);
-            fn enable_input(&mut self, on: bool, _internal: private::Internal);
-            fn enable_input_in_sleep_mode(&mut self, on: bool, _internal: private::Internal);
-            fn is_input_high(&self, _internal: private::Internal) -> bool;
-            fn connect_input_to_peripheral(&mut self, signal: InputSignal, _internal: private::Internal);
-            fn connect_input_to_peripheral_with_options(
-                &mut self,
-                signal: InputSignal,
-                invert: bool,
-                force_via_gpio_mux: bool,
-                _internal: private::Internal,
-            );
-            fn disconnect_input_from_peripheral(&mut self, signal: InputSignal, _internal: private::Internal);
         }
     }
 }